Refactor panels to its own directory
This is primarily done to unclutter the root directory
