+++ web/img-mngr/src/app.html
@@ -1,8 +1,8 @@
 <template>
   <require from="bootstrap/css/bootstrap.css"></require>
   <require from="./styles.css"></require>
-  <require from="./download"></require>
-  <require from="./breadcrumbs"></require>
+  <require from="./components/panels/download"></require>
+  <require from="./components/panels/breadcrumbs"></require>
 
   <div id="rootContainer" class="container">
     <div class="floater header">
